# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "battery life — API для управления заявками расчёта времени работы аккумуляторов",
        "title": "battery life API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/battery_life/all-battery_life": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает заявки пользователя, а для модератора - все заявки. Поддерживает фильтрацию по датам и статусу.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Получить список заявок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начальная дата (YYYY-MM-DD)",
                        "name": "from-date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (YYYY-MM-DD)",
                        "name": "to-date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Статус заявки",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/battery_life/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию о заявке и ее позициях.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Получить заявку по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life/{id}/delete-battery_life": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выполняет логическое удаление черновика заявки.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Удалить заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life/{id}/edit-battery_life": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет данные черновика заявки.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Изменить заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные заявки",
                        "name": "battery_life",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life/{id}/finish-battery_life": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет статус заявки на `completed` или `rejected`. Доступно только модератору.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Завершить заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус заявки",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.StatusJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life/{id}/form-battery_life": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Переводит черновик в статус `formed`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Сформировать заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life_item/add/{battery_life_type_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет аккумулятор в черновик пользователя. При отсутствии черновика он создается.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Добавить аккумулятор в заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID типа аккумулятора",
                        "name": "battery_life_type_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Параметры позиции",
                        "name": "item",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeItemJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryLifeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/battery_life_type/create-battery_life_type": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый тип аккумулятора. Доступно авторизованному пользователю.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Создать аккумулятор",
                "parameters": [
                    {
                        "description": "Данные аккумулятора",
                        "name": "battery_life_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/battery_life_type/{id}": {
            "get": {
                "description": "Возвращает данные одного аккумулятора.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery life"
                ],
                "summary": "Получить аккумулятор по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID аккумулятора",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RIP2026_internal_app_serializer.BatteryTypeJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "RIP2026_internal_app_serializer.BatteryLifeItemJSON": {
            "type": "object",
            "properties": {
                "battery_life_id": {
                    "type": "integer"
                },
                "battery_type_id": {
                    "type": "integer"
                },
                "current_ma": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "runtime_hours": {
                    "type": "number"
                }
            }
        },
        "RIP2026_internal_app_serializer.BatteryLifeJSON": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "completed_item_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "formed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator_login": {
                    "description": "опционально, как в SystemLoadToJSON",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_runtime_hours": {
                    "type": "number"
                }
            }
        },
...
        "RIP2026_internal_app_serializer.StatusJSON": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}